+++ builder/test-progs/multi/module-4.cpp
@@ -1,5 +1,5 @@
 #include "module-4.hpp"
 
-unsigned long func_4(unsigned long in) {
+u32 func_4(u32 in) {
   return in % 269;
 }
